Separate node code to its own file, remove some deadcode
